added appbar to custom widget
