+++ lib/ui/components/custom_app_bar_widget.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+import 'package:hotel_app/resources/resources.dart';
+
+class CustomAppBarWidget extends StatelessWidget implements PreferredSizeWidget {
+  const CustomAppBarWidget({
+    super.key,
+    required this.title,
+  });
+
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      title: Text(
+        title,
+        style: const TextStyle(
+          color: Colors.black,
+          fontSize: 18,
+          fontWeight: FontWeight.w500,
+        ),
+      ),
+      centerTitle: true,
+      leading: IconButton(
+        icon: Image.asset(AppImages.backwardArrow),
+        onPressed: () {
+          Navigator.pop(context);
+        },
+      ),
+      scrolledUnderElevation: 0,
+    );
+  }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(60);
+}
