+++ lib/ui/widgets/Hotel/hotel_widget.dart
@@ -67,6 +67,7 @@ class _HotelWidgetState extends State<HotelWidget> {
           ),
         ),
         centerTitle: true,
+        scrolledUnderElevation: 0,
       ),
       body: ListView(
         children: [
