+++ lib/ui/widgets/OrderPaid/order_paid_widget.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hotel_app/resources/app_colors.dart';
 import 'package:hotel_app/resources/resources.dart';
+import 'package:hotel_app/ui/components/custom_app_bar_widget.dart';
 import 'package:hotel_app/ui/components/custom_bottom_navigation_bar.dart';
 
 class OrderPaidWidget extends StatelessWidget {
@@ -45,35 +46,3 @@ class OrderPaidWidget extends StatelessWidget {
   }
 }
 
-class CustomAppBarWidget extends StatelessWidget implements PreferredSizeWidget {
-  const CustomAppBarWidget({
-    super.key,
-    required this.title,
-  });
-
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return AppBar(
-      title: Text(
-        title,
-        style: const TextStyle(
-          color: Colors.black,
-          fontSize: 18,
-          fontWeight: FontWeight.w500,
-        ),
-      ),
-      centerTitle: true,
-      leading: IconButton(
-        icon: Image.asset(AppImages.backwardArrow),
-        onPressed: () {
-          Navigator.pop(context);
-        },
-      ),
-    );
-  }
-
-  @override
-  Size get preferredSize => const Size.fromHeight(101);
-}
